[New] expose createDynamicImportTransform and getImportSource
While working on babel/babel#9552, I realized I needed two functions:
- createDynamicImportTransform is basically all what this plugin
  does, but it can be invoked directly.
- getImportSource is needed to re-use the same import argument
  stringifying logic for the AMD and SystemJS transforms
